Removed unnecessary functions.  Renamed objects_operation to object_operations. Renamed image_parser to image_operations.
